pyproject.toml
[project]
name = "sa-assistant"
version = "0.1.0"
description = "AI assistant for StackAdapt"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.1",
    "asana>=5.1.0",
    "chromadb>=1.0.15",
    "dotenv>=0.9.9",
    "flake8-pyproject>=1.2.3",
    "google-api-python-client>=2.170.0",
    "google-auth>=2.40.2",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-genai>=1.17.0",
    "jira>=3.8.0",
    "mcp[cli]>=1.9.0",
    "openai>=1.79.0",
    "openai-agents>=0.0.15",
    "pydantic>=2.11.4",
    "python-lsp-server[all]>=1.12.2",
    "pytz>=2025.2",
    "pyyaml>=6.0.2",
    "slack-sdk>=3.35.0",
    "sqlmodel>=0.0.24",
]
[tool.flake8]
max-line-length = 200
[tool.pylint]
disable = ["missing-module-docstring",
           "missing-class-docstring",
           "missing-function-docstring"
]
